requirements.txt
sentence-transformers
hnswlib
pandas
openai
python-dotenv
sqlalchemy
psycopg2-binary
psycopg2
Flask
instructor
pydantic
anthropic
elasticsearch